Refactor module and rename TaggedMarker -> Head
