+++ cbor/src/decode/mod.rs
@@ -20,9 +20,9 @@ There are three primary ways to use the decoder:
     stream piece by piece. This approach gives you fine-grained control and
     avoids intermediate allocations.
 
-3.  **Low-level peek with [`TaggedMarker`]:** For optimised parsing where
+3.  **Low-level peek with [`Head`]:** For optimised parsing where
     you want to dispatch on type and length without using a closure or
-    materialising indefinite-length contents, parse a [`TaggedMarker`] and
+    materialising indefinite-length contents, parse a [`Head`] and
     match on its [`Marker`] directly. The caller chooses whether to walk
     the contents of arrays, maps, and chunked strings.
 
@@ -88,6 +88,9 @@ use super::*;
 use core::{ops::Range, str::Utf8Error};
 use num_traits::{FromPrimitive, ToPrimitive};
 use thiserror::Error;
+
+mod head;
+mod series;
 
 /// An error that can occur during CBOR decoding.
 #[derive(Error, Debug)]
@@ -166,13 +169,15 @@ pub trait FromCbor: Sized {
 }
 
 /// A type alias for a generic, untyped CBOR sequence.
-pub type Sequence<'a> = super::decode_seq::Series<'a, 0>;
+pub type Sequence<'a> = series::Series<'a, 0>;
 /// A type alias for a [`Series`] that represents a CBOR array.
-pub type Array<'a> = super::decode_seq::Series<'a, 1>;
+pub type Array<'a> = series::Series<'a, 1>;
 /// A type alias for a [`Series`] that represents a CBOR map.
-pub type Map<'a> = super::decode_seq::Series<'a, 2>;
+pub type Map<'a> = series::Series<'a, 2>;
+/// The head of a single CBOR data item.
+pub use head::{Head, Marker};
 /// A stateful iterator for decoding a sequence of CBOR items (e.g., an array or map).
-pub use super::decode_seq::Series;
+pub use series::Series;
 
 /// Represents a single, decoded CBOR data item.
 pub enum Value<'a, 'b: 'a> {
@@ -282,364 +287,9 @@ impl<'a, 'b: 'a> core::fmt::Debug for Value<'a, 'b> {
     }
 }
 
-fn parse_tags(data: &[u8]) -> Result<(Vec<u64>, bool, usize), Error> {
-    let mut tags = Vec::new();
-    let mut offset = 0;
-    let mut shortest = true;
-
-    while let Some(marker) = data.get(offset) {
-        match (marker >> 5, marker & 0x1F) {
-            (6, minor) => {
-                offset += 1;
-                let (tag, s, o) = parse_uint_minor(minor, &data[offset..])?;
-                tags.push(tag);
-                shortest &= s;
-                offset = offset.checked_add(o).ok_or(Error::TooBig)?;
-            }
-            _ => break,
-        }
-    }
-    Ok((tags, shortest, offset))
-}
-
-#[inline]
-fn to_array<const N: usize>(data: &[u8]) -> Result<[u8; N], Error> {
-    match data.len().cmp(&N) {
-        core::cmp::Ordering::Less => Err(Error::NeedMoreData(N - data.len())),
-        core::cmp::Ordering::Equal => Ok(data.try_into().unwrap()),
-        core::cmp::Ordering::Greater => Ok(data[0..N].try_into().unwrap()),
-    }
-}
-
-#[inline]
-fn parse_uint_minor(minor: u8, data: &[u8]) -> Result<(u64, bool, usize), Error> {
-    match minor {
-        24 => {
-            if let Some(val) = data.first() {
-                Ok((*val as u64, *val > 23, 1))
-            } else {
-                Err(Error::NeedMoreData(1))
-            }
-        }
-        25 => {
-            let v = u16::from_be_bytes(to_array(data)?);
-            Ok((v as u64, v > u8::MAX as u16, 2))
-        }
-        26 => {
-            let v = u32::from_be_bytes(to_array(data)?);
-            Ok((v as u64, v > u16::MAX as u32, 4))
-        }
-        27 => {
-            let v = u64::from_be_bytes(to_array(data)?);
-            Ok((v, v > u32::MAX as u64, 8))
-        }
-        val if val < 24 => Ok((val as u64, true, 0)),
-        _ => Err(Error::InvalidMinorValue(minor)),
-    }
-}
-
-/// The head of a single CBOR data item.
-///
-/// `Marker` captures the CBOR major type and the value carried directly in
-/// the type marker encoding — a scalar, a definite-length payload length,
-/// or an element count. It is the payload returned by the [`TaggedMarker`]
-/// [`FromCbor`] implementation when you only need to dispatch on type
-/// without paying for a full decode.
-///
-/// # What this *does not* tell you
-///
-/// `Marker` does **not** carry the byte length of the encoded CBOR item as
-/// a whole. The values inside the variants describe the item itself, not
-/// its encoded size:
-///
-/// - [`Array(Some(count))`][Self::Array] / [`Map(Some(count))`][Self::Map]
-///   carry the **element count** — for a map, the number of key-value
-///   pairs. They say nothing about how many bytes the contained items
-///   occupy.
-/// - [`Bytes(Some(len))`][Self::Bytes] / [`Text(Some(len))`][Self::Text]
-///   carry the **payload length in bytes**. The payload itself sits in
-///   the buffer immediately after the marker head; it is not consumed
-///   by the marker decode.
-/// - The `None` variants of `Bytes`, `Text`, `Array`, and `Map` carry no
-///   length information at all — the contents are indefinite-length and
-///   must be walked to a break byte.
-///
-/// `Marker` derives [`Debug`], [`Clone`], and [`PartialEq`]. `PartialEq`
-/// follows IEEE-754 semantics for [`Float`][Self::Float] (`NaN != NaN`).
-///
-/// # Bytes consumed
-///
-/// The byte count returned alongside a `Marker` by [`FromCbor`] covers
-/// the encoding of the type marker itself — including any length
-/// prefix — but never the variable-length payload that follows:
-///
-/// - **Scalars** (integers, floats, booleans, null, undefined, simple
-///   values): the full encoding is consumed.
-/// - **Definite-length strings** ([`Bytes(Some(_))`][Self::Bytes],
-///   [`Text(Some(_))`][Self::Text]): only the head byte and length
-///   prefix are consumed; the payload bytes remain in the buffer and
-///   the `Some(len)` value gives their length.
-/// - **Indefinite-length strings** ([`Bytes(None)`][Self::Bytes],
-///   [`Text(None)`][Self::Text]): only the single head byte is consumed;
-///   the chunks and the trailing break byte remain in the buffer.
-/// - **Arrays and maps** ([`Array`][Self::Array], [`Map`][Self::Map],
-///   either `Some` or `None`): only the head byte and (for definite
-///   collections) the length prefix are consumed; the contained items
-///   remain in the buffer for the caller to walk.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Marker {
-    /// An unsigned integer (CBOR major type 0).
-    UnsignedInteger(u64),
-    /// A negative integer (CBOR major type 1), stored as the raw value `n` where the actual value is `-1 - n`.
-    NegativeInteger(u64),
-    /// A byte string (CBOR major type 2). `Some(len)` is the payload
-    /// length in bytes — the payload itself remains in the buffer
-    /// immediately after the marker; `None` indicates an indefinite-length
-    /// string whose chunks are still in the buffer awaiting parsing.
-    Bytes(Option<u64>),
-    /// A text string (CBOR major type 3). `Some(len)` is the payload
-    /// length in bytes — the payload itself remains in the buffer
-    /// immediately after the marker; `None` indicates an indefinite-length
-    /// string whose chunks are still in the buffer awaiting parsing.
-    Text(Option<u64>),
-    /// A CBOR array (major type 4). `Some(count)` is the number of
-    /// elements for definite-length arrays — not a byte length;
-    /// `None` indicates an indefinite-length array whose elements are
-    /// still in the buffer, terminated by a break byte.
-    Array(Option<u64>),
-    /// A CBOR map (major type 5). `Some(count)` is the number of
-    /// key-value pairs for definite-length maps — not a byte length;
-    /// `None` indicates an indefinite-length map whose pairs are still in
-    /// the buffer, terminated by a break byte.
-    Map(Option<u64>),
-    /// The boolean value `false` (CBOR simple value 20).
-    False,
-    /// The boolean value `true` (CBOR simple value 21).
-    True,
-    /// The null value (CBOR simple value 22).
-    Null,
-    /// The undefined value (CBOR simple value 23).
-    Undefined,
-    /// An unassigned simple value (CBOR simple values 0–19, 24–31).
-    Simple(u8),
-    /// A floating-point value (CBOR major type 7).
-    Float(f64),
-    /// An 'end' marker for indefinite length items
-    End,
-}
-
-/// A [`Marker`] preceded by zero or more CBOR semantic tags.
-///
-/// `TaggedMarker` is the head of a tagged or untagged CBOR item produced by
-/// the [`FromCbor`] implementation. It is the entry point for low-level
-/// parsing, in contrast to the closure-driven [`parse_value`] family.
-///
-/// Use [`parse::<TaggedMarker>(data)`][parse] to peek at the next item, or
-/// [`parse::<(TaggedMarker, bool, usize)>(data)`][parse] when you also need
-/// the canonical-encoding flag and the byte count consumed by the marker
-/// itself (see [`Marker`] for the consumption rules — this count is *not*
-/// the size of the encoded item for arrays, maps, or indefinite-length
-/// strings).
-///
-/// # When to use this over [`parse_value`]
-///
-/// - **No closure required.** Match directly on the returned [`Marker`]
-///   instead of threading control flow through an `FnOnce`, which avoids
-///   borrow-checker friction and lets the caller propagate any error type.
-/// - **No contiguous materialisation.** [`parse_value`] eagerly collects
-///   indefinite-length string chunks into a `Vec<Range<usize>>` and
-///   constructs nested [`Series`] iterators for arrays and maps; with
-///   `TaggedMarker` the caller decides whether to walk the chunks or
-///   sub-items at all, and may skip them byte-wise instead of parsing
-///   them.
-pub struct TaggedMarker {
-    /// CBOR major-type-6 tags preceding the item, in encoding order. Empty
-    /// if the item is untagged.
-    pub tags: Vec<u64>,
-    /// The decoded marker for the item itself.
-    pub marker: Marker,
-}
-
-impl core::fmt::Display for TaggedMarker {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        let prefix = if self.tags.is_empty() {
-            "Untagged"
-        } else {
-            "Tagged"
-        };
-        match self.marker {
-            Marker::UnsignedInteger(_) => write!(f, "{prefix} Unsigned Integer"),
-            Marker::NegativeInteger(_) => write!(f, "{prefix} Negative Integer"),
-            Marker::Bytes(Some(_)) => write!(f, "{prefix} Definite-length Byte String"),
-            Marker::Bytes(None) => write!(f, "{prefix} Indefinite-length Byte String"),
-            Marker::Text(Some(_)) => write!(f, "{prefix} Definite-length Text String"),
-            Marker::Text(None) => write!(f, "{prefix} Indefinite-length Text String"),
-            Marker::Array(Some(_)) => write!(f, "{prefix} Definite-length Array"),
-            Marker::Array(None) => write!(f, "{prefix} Indefinite-length Array"),
-            Marker::Map(Some(_)) => write!(f, "{prefix} Definite-length Map"),
-            Marker::Map(None) => write!(f, "{prefix} Indefinite-length Map"),
-            Marker::False => write!(f, "{prefix} False"),
-            Marker::True => write!(f, "{prefix} True"),
-            Marker::Null => write!(f, "{prefix} Null"),
-            Marker::Undefined => write!(f, "{prefix} Undefined"),
-            Marker::Simple(v) => write!(f, "{prefix} Simple Value {v}"),
-            Marker::Float(_) => write!(f, "{prefix} Float"),
-            Marker::End => write!(f, "End of sequence marker"),
-        }
-    }
-}
-
-impl FromCbor for TaggedMarker {
-    type Error = Error;
-
-    fn from_cbor(data: &[u8]) -> Result<(Self, bool, usize), Self::Error> {
-        let (tags, mut shortest, mut offset) = parse_tags(data)?;
-        let Some(marker) = data.get(offset) else {
-            return Err(Error::NeedMoreData(1));
-        };
-        offset += 1;
-        let data = &data[offset..];
-
-        let (marker, shortest, len) = match (marker >> 5, marker & 0x1F) {
-            (0, minor) => parse_uint_minor(minor, data)
-                .map(|(v, s, len)| (Marker::UnsignedInteger(v), shortest && s, len))?,
-            (1, minor) => parse_uint_minor(minor, data)
-                .map(|(v, s, len)| (Marker::NegativeInteger(v), shortest && s, len))?,
-            (2, 31) => (Marker::Bytes(None), shortest, 0),
-            (2, minor) => {
-                /* Known length byte string */
-                parse_uint_minor(minor, data)
-                    .map(|(v, s, len)| (Marker::Bytes(Some(v)), shortest && s, len))?
-            }
-            (3, 31) => {
-                /* Indefinite length text string */
-                (Marker::Text(None), shortest, 0)
-            }
-            (3, minor) => {
-                /* Known length text string */
-                parse_uint_minor(minor, data)
-                    .map(|(v, s, len)| (Marker::Text(Some(v)), shortest && s, len))?
-            }
-            (4, 31) => {
-                /* Indefinite length array */
-                (Marker::Array(None), shortest, 0)
-            }
-            (4, minor) => {
-                /* Known length array */
-                parse_uint_minor(minor, data)
-                    .map(|(v, s, len)| (Marker::Array(Some(v)), shortest && s, len))?
-            }
-            (5, 31) => {
-                /* Indefinite length map */
-                (Marker::Map(None), shortest, 0)
-            }
-            (5, minor) => {
-                /* Known length map */
-                parse_uint_minor(minor, data)
-                    .map(|(v, s, len)| (Marker::Map(Some(v)), shortest && s, len))?
-            }
-            (6, _) => unreachable!("CBOR major type 6 (tags) consumed before dispatch"),
-            (7, 20) => {
-                /* False */
-                (Marker::False, shortest, 0)
-            }
-            (7, 21) => {
-                /* True */
-                (Marker::True, shortest, 0)
-            }
-            (7, 22) => {
-                /* Null */
-                (Marker::Null, shortest, 0)
-            }
-            (7, 23) => {
-                /* Undefined */
-                (Marker::Undefined, shortest, 0)
-            }
-            (7, minor @ 0..=19) => {
-                /* Unassigned simple type */
-                (Marker::Simple(minor), shortest, 0)
-            }
-            (7, 24) => {
-                /* Unassigned simple type */
-                let Some(v) = data.first() else {
-                    return Err(Error::NeedMoreData(1));
-                };
-                if *v < 32 {
-                    return Err(Error::InvalidSimpleType(*v));
-                }
-                (Marker::Simple(*v), shortest, 1)
-            }
-            (7, 25) => {
-                /* FP16 */
-                let v = half::f16::from_be_bytes(to_array(data)?);
-                (Marker::Float(v.into()), shortest, 2)
-            }
-            (7, 26) => {
-                /* FP32 */
-                let v = f32::from_be_bytes(to_array(data)?);
-                if shortest {
-                    match v.classify() {
-                        core::num::FpCategory::Nan
-                        | core::num::FpCategory::Infinite
-                        | core::num::FpCategory::Zero => {
-                            // There is an FP16 representation that is shorter
-                            shortest = false;
-                        }
-                        core::num::FpCategory::Subnormal | core::num::FpCategory::Normal => {
-                            if let Some(v16) = <half::f16 as num_traits::FromPrimitive>::from_f32(v)
-                                && <half::f16 as num_traits::ToPrimitive>::to_f32(&v16) == Some(v)
-                            {
-                                shortest = false;
-                            }
-                        }
-                    }
-                }
-                (Marker::Float(v.into()), shortest, 4)
-            }
-            (7, 27) => {
-                /* FP64 */
-                let v = f64::from_be_bytes(to_array(data)?);
-                if shortest {
-                    match v.classify() {
-                        core::num::FpCategory::Nan
-                        | core::num::FpCategory::Infinite
-                        | core::num::FpCategory::Zero => {
-                            // There is an FP16 representation that is shorter
-                            shortest = false;
-                        }
-                        core::num::FpCategory::Subnormal | core::num::FpCategory::Normal => {
-                            if let Some(v32) = f32::from_f64(v) {
-                                if v32.to_f64() == Some(v) {
-                                    shortest = false;
-                                }
-                            } else if let Some(v16) =
-                                <half::f16 as num_traits::FromPrimitive>::from_f64(v)
-                                && <half::f16 as num_traits::ToPrimitive>::to_f64(&v16) == Some(v)
-                            {
-                                shortest = false;
-                            }
-                        }
-                    }
-                }
-                (Marker::Float(v), shortest, 8)
-            }
-            // A break code is only a valid End marker when untagged.
-            // `offset == 1` means no tag bytes preceded it (offset = tag
-            // bytes consumed + 1 for the marker byte). A tagged break
-            // falls through to the catch-all and is rejected.
-            (7, 31) if offset == 1 => (Marker::End, true, 0),
-            (7, minor) => {
-                return Err(Error::InvalidSimpleType(minor));
-            }
-            _ => unreachable!("CBOR major type is 3 bits, all values 0-7 handled above"),
-        };
-        Ok((TaggedMarker { tags, marker }, shortest, offset + len))
-    }
-}
-
 /// Skips over the next CBOR data item without materialising its contents.
 ///
-/// Walks the encoding using [`TaggedMarker`] dispatch. Indefinite-length
+/// Walks the encoding using [`Head`] dispatch. Indefinite-length
 /// string chunks are walked in place with no allocation; arrays and maps
 /// are drained through a stack-only [`Series`] cursor — no `Vec` for
 /// chunk lists or nested values is constructed.
@@ -658,14 +308,14 @@ impl FromCbor for TaggedMarker {
 ///
 /// [RFC 8949 §4.2.1]: https://www.rfc-editor.org/rfc/rfc8949.html#section-4.2.1
 pub fn skip_value(data: &[u8], mut max_recursion: usize) -> Result<(bool, usize), Error> {
-    let (marker, mut shortest, mut offset) = parse::<(TaggedMarker, bool, usize)>(data)?;
+    let (marker, mut shortest, mut offset) = parse::<(Head, bool, usize)>(data)?;
     match marker.marker {
         Marker::Bytes(Some(v)) | Marker::Text(Some(v)) => {
             Ok((shortest, checked_offset(offset, v)?))
         }
         Marker::Bytes(None) => loop {
             let (inner_marker, inner_shortest, len) =
-                parse::<(TaggedMarker, bool, usize)>(&data[offset..])?;
+                parse::<(Head, bool, usize)>(&data[offset..])?;
             offset = offset.checked_add(len).ok_or(Error::TooBig)?;
 
             match inner_marker.marker {
@@ -673,13 +323,13 @@ pub fn skip_value(data: &[u8], mut max_recursion: usize) -> Result<(bool, usize)
                     offset = checked_offset(offset, len)?;
                     shortest &= inner_shortest;
                 }
-                Marker::End => return Ok((shortest, offset)),
+                Marker::Break => return Ok((shortest, offset)),
                 _ => return Err(Error::InvalidChunk),
             }
         },
         Marker::Text(None) => loop {
             let (inner_marker, inner_shortest, len) =
-                parse::<(TaggedMarker, bool, usize)>(&data[offset..])?;
+                parse::<(Head, bool, usize)>(&data[offset..])?;
             offset = offset.checked_add(len).ok_or(Error::TooBig)?;
 
             match inner_marker.marker {
@@ -687,7 +337,7 @@ pub fn skip_value(data: &[u8], mut max_recursion: usize) -> Result<(bool, usize)
                     offset = checked_offset(offset, len)?;
                     shortest &= inner_shortest;
                 }
-                Marker::End => return Ok((shortest, offset)),
+                Marker::Break => return Ok((shortest, offset)),
                 _ => return Err(Error::InvalidChunk),
             }
         },
@@ -737,7 +387,7 @@ fn offset_extent(data: &[u8], offset: usize, len: u64) -> Result<Range<usize>, E
 /// fully-formed [`Value`] to `f`. Indefinite-length strings are gathered
 /// into chunk lists and arrays/maps are presented as nested [`Series`]
 /// iterators, so the closure sees a uniform view regardless of encoding.
-/// For finer control without a closure, parse a [`TaggedMarker`] instead.
+/// For finer control without a closure, parse a [`Head`] instead.
 ///
 /// On success, returns the result of `f` and the total number of bytes
 /// consumed from the input slice.
@@ -746,7 +396,7 @@ where
     F: FnOnce(Value, bool, &[u64]) -> Result<T, E>,
     E: From<Error>,
 {
-    let (marker, mut shortest, mut offset) = parse::<(TaggedMarker, bool, usize)>(data)?;
+    let (marker, mut shortest, mut offset) = parse::<(Head, bool, usize)>(data)?;
     match marker.marker {
         Marker::UnsignedInteger(v) => f(Value::UnsignedInteger(v), shortest, &marker.tags),
         Marker::NegativeInteger(v) => f(Value::NegativeInteger(v), shortest, &marker.tags),
@@ -759,7 +409,7 @@ where
             let mut chunks = Vec::new();
             loop {
                 let (inner_marker, inner_shortest, len) =
-                    parse::<(TaggedMarker, bool, usize)>(&data[offset..])?;
+                    parse::<(Head, bool, usize)>(&data[offset..])?;
 
                 offset = offset.checked_add(len).ok_or(Error::TooBig)?;
 
@@ -770,7 +420,7 @@ where
                         chunks.push(extent);
                         shortest &= inner_shortest;
                     }
-                    Marker::End => {
+                    Marker::Break => {
                         break f(Value::ByteStream(chunks), shortest, &marker.tags);
                     }
                     _ => {
@@ -792,7 +442,7 @@ where
             let mut chunks = Vec::new();
             loop {
                 let (inner_marker, inner_shortest, len) =
-                    parse::<(TaggedMarker, bool, usize)>(&data[offset..])?;
+                    parse::<(Head, bool, usize)>(&data[offset..])?;
 
                 offset = offset.checked_add(len).ok_or(Error::TooBig)?;
 
@@ -803,7 +453,7 @@ where
                         chunks.push(core::str::from_utf8(&data[extent]).map_err(Into::into)?);
                         shortest &= inner_shortest;
                     }
-                    Marker::End => {
+                    Marker::Break => {
                         break f(Value::TextStream(&chunks), shortest, &marker.tags);
                     }
                     _ => {
@@ -828,7 +478,7 @@ where
         Marker::Undefined => f(Value::Undefined, shortest, &marker.tags),
         Marker::Simple(v) => f(Value::Simple(v), shortest, &marker.tags),
         Marker::Float(v) => f(Value::Float(v), shortest, &marker.tags),
-        Marker::End => return Err(Error::InvalidSimpleType(31).into()),
+        Marker::Break => return Err(Error::InvalidSimpleType(31).into()),
     }
     .map(|r| (r, offset))
 }
@@ -860,7 +510,7 @@ where
     F: FnOnce(&mut Array, bool, &[u64]) -> Result<T, E>,
     E: From<Error>,
 {
-    let (marker, shortest, mut offset) = parse::<(TaggedMarker, bool, usize)>(data)?;
+    let (marker, shortest, mut offset) = parse::<(Head, bool, usize)>(data)?;
     match marker.marker {
         Marker::Array(count) => {
             let mut a = Array::try_new(data, count, &mut offset)?;
@@ -881,7 +531,7 @@ where
     F: FnOnce(&mut Map, bool, &[u64]) -> Result<T, E>,
     E: From<Error>,
 {
-    let (marker, shortest, mut offset) = parse::<(TaggedMarker, bool, usize)>(data)?;
+    let (marker, shortest, mut offset) = parse::<(Head, bool, usize)>(data)?;
     match marker.marker {
         Marker::Map(count) => {
             let mut m = Map::try_new(data, count, &mut offset)?;
@@ -929,7 +579,7 @@ impl FromCbor for u64 {
 
     #[inline]
     fn from_cbor(data: &[u8]) -> Result<(Self, bool, usize), Self::Error> {
-        let (marker, shortest, offset) = parse::<(TaggedMarker, bool, usize)>(data)?;
+        let (marker, shortest, offset) = parse::<(Head, bool, usize)>(data)?;
         if let Marker::UnsignedInteger(v) = marker.marker {
             Ok((v, shortest && marker.tags.is_empty(), offset))
         } else {
@@ -964,7 +614,7 @@ impl FromCbor for i64 {
 
     #[inline]
     fn from_cbor(data: &[u8]) -> Result<(Self, bool, usize), Self::Error> {
-        let (marker, shortest, offset) = parse::<(TaggedMarker, bool, usize)>(data)?;
+        let (marker, shortest, offset) = parse::<(Head, bool, usize)>(data)?;
         match marker.marker {
             Marker::UnsignedInteger(v) => Ok((
                 i64::try_from(v)?,
@@ -1016,7 +666,7 @@ impl FromCbor for f64 {
 
     #[inline]
     fn from_cbor(data: &[u8]) -> Result<(Self, bool, usize), Self::Error> {
-        let (marker, shortest, offset) = parse::<(TaggedMarker, bool, usize)>(data)?;
+        let (marker, shortest, offset) = parse::<(Head, bool, usize)>(data)?;
         if let Marker::Float(v) = marker.marker {
             Ok((v, shortest && marker.tags.is_empty(), offset))
         } else {
@@ -1033,7 +683,7 @@ impl FromCbor for bool {
 
     #[inline]
     fn from_cbor(data: &[u8]) -> Result<(Self, bool, usize), Self::Error> {
-        let (marker, shortest, offset) = parse::<(TaggedMarker, bool, usize)>(data)?;
+        let (marker, shortest, offset) = parse::<(Head, bool, usize)>(data)?;
         match marker.marker {
             Marker::False => Ok((false, shortest && marker.tags.is_empty(), offset)),
             Marker::True => Ok((true, shortest && marker.tags.is_empty(), offset)),
